This PR includes the work in #20900 and supercedes it.

I tested this this a variable product which had 500 variations and 4 different attributes.

The original PR queried variation IDs then passed or variation IDs to the 2nd mysql query using IN clause. These 2 queries took approx 2.5ms total running direct SQL queries to test. I was slightly concerned about how many IDs could be passed at once so I looked into a JOIN solution instead and tried to get the join < 2.5ms without a full table scan so it was still performant.

I switched it to a subquery which on average was quicker, and also swapped the LIKE for explicit attribute names. @kloon @rodrigoprimo Feedback welcome on this.

On the product page with my test product, using the network tab in browser tools the response times went from around 2s -> 1.5s after this patch. From what I can tell, the logic still works like before.

To test, create a variable product with many variations (>30) and use the network inspector to see how faster the get_variation ajax call is.

> Improved speed of the `find_matching_product_variation` variation lookup function.